feat: migrate controller to ts
